ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.4",
    "google-cloud-aiplatform>=1.136.0",
    "google-cloud-vision>=3.12.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "google-genai",
    "google-adk",
    "google-cloud-storage>=3.9.0",
    "ultralytics>=8.4.13",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.0",
    "huggingface-hub>=1.4.1",
]
